Add interpreter

+++ src/parsing/tokens.rs
@@ -1,6 +1,6 @@
 #[derive(Debug)]
 pub enum Token {
-    ADD(i64),
+    ADD(i8),
     MOVE(i64),
     LOOP(Vec<Token>),
     PRINT,
